refactor: consolidate lockfile ignore logic and identification into util package
- Centralized lockfile list and logic in util/lockfiles.go.
- Replaced redundant IsLockFile and appendLockFileExcludes implementations.
- Optimized IsLockFile for zero-allocation suffix checking.
- Updated all call sites and tests to use the shared utility.
- Added util/AGENTS.md with mandatory path trimming guidelines.
- Updated Bazel BUILD files.
