+++ core/config/BUILD.bazel
@@ -8,7 +8,7 @@ go_library(
     deps = [
         "//core/prompts",
         "//llm",
-        "//tools",
+        "//util",
         "@com_github_spf13_viper//:viper",
     ],
 )
