+++ util/BUILD.bazel
@@ -2,13 +2,19 @@ load("@rules_go//go:def.bzl", "go_library", "go_test")
 
 go_library(
     name = "util",
-    srcs = ["fs.go"],
+    srcs = [
+        "fs.go",
+        "lockfiles.go",
+    ],
     importpath = "github.com/menny/cassandra/util",
     visibility = ["//visibility:public"],
 )
 
 go_test(
     name = "util_test",
-    srcs = ["fs_test.go"],
+    srcs = [
+        "fs_test.go",
+        "lockfiles_test.go",
+    ],
     embed = [":util"],
 )
